added time storage and health decay

+++ src/TextureManager.h
@@ -8,6 +8,6 @@ class TextureManager {
         static void drawBG(SDL_Texture* tex);
         static SDL_Texture* loadTextureFromText(const char* text, const char* fontFile, int size, SDL_Color color);
         static void drawText(SDL_Texture* text, SDL_Rect dest);
-        static SDL_Texture* progressBar(int progress, int maxProgress, SDL_Color bg_color, SDL_Color fg_color);
+        static SDL_Texture* progressBar(float progress, int maxProgress, SDL_Color bg_color, SDL_Color fg_color);
         static void drawProgressBar(SDL_Texture* tex, SDL_Rect dest);
 };
